fix: Sending of the dynamic table size update
Advertise the selected dynamic table size if the client's table size is
not equal to Tempesta FW dynamic table size.

Details: when Tempesta receives SETTINGS frame and
SETTINGS_HEADER_TABLE_SIZE in this frame is greater than default
Tempesta's 4096, Tempesta advertise the size that it will be used.
Before this patch Tempesta ignored it when Tempesta's table size is
equal to default 4096 and size advertised by the client is greater than
4096. In other cases Tempesta did send updates.

For reference see: RFC7541 Section 4.2

+++ fw/hpack.h
@@ -305,7 +305,7 @@ int tfw_hpack_transform(TfwHttpResp *__restrict resp, TfwStr *__restrict hdr);
 int tfw_hpack_encode(TfwHttpResp *__restrict resp, TfwStr *__restrict hdr,
 		     bool use_pool, bool dyn_indexing);
 void tfw_hpack_set_rbuf_size(TfwHPackETbl *__restrict tbl,
-			     unsigned short new_size);
+			     unsigned int new_size);
 int tfw_hpack_decode(TfwHPack *__restrict hp, unsigned char *__restrict src,
 		     unsigned long n, TfwHttpReq *__restrict req,
 		     unsigned int *__restrict parsed);
